+++ src/main/notificationtemplates/purchasenotification.go
@@ -6,14 +6,14 @@ import (
 	krakenapi "github.com/beldur/kraken-go-api-client"
 )
 
-func NewPurchaseNotification(pair string, amount float32, orderPrice float32, transactionId string, completedOrder krakenapi.Order) PurchaseNotification {
+func NewPurchaseNotification(pair string, amount float64, orderPrice float64, transactionId string, completedOrder krakenapi.Order) PurchaseNotification {
 	return PurchaseNotification{pair: pair, amount: amount, orderPrice: orderPrice, transactionId: transactionId, completedOrder: completedOrder}
 }
 
 type PurchaseNotification struct {
 	pair           string
-	amount         float32
-	orderPrice     float32
+	amount         float64
+	orderPrice     float64
 	transactionId  string
 	completedOrder krakenapi.Order // TODO Take individual values
 }
